+++ src/convert/index.ts
@@ -20,4 +20,6 @@ export * from './convert-encounter';
 export * from './convert-related-person';
 export * from './convert-coverage';
 export * from './convert-clinical-impression';
+export * from './convert-consent';
+export * from './convert-practitioner-role';
 export * from './schemaorg-to-gaia-x';